LF-12564 wlwindow: fix the issue of video hang after resizing by mouse
When resizing video by mouse, there are two threads (render/display thread)
that may wait for each other to release the mutex (render_lock/sync_mutex)
they hold, which will cause the video pipeline to hang.

Due to the update of gstreamer, the 'surface_commit' in the render thread
is migrated to the display thread, so delete the render_lock in the display
thread, so that the render thread does not need to wait for the display thread.

Signed-off-by: Chao Guo <chao.guo@nxp.com>