webrtcnice: Wait for completion of proxy and mdns resolve tasks
This commit ensures all outstanding name resolving for mdns and proxy is
finished before the nice agent thread is shut down. This is important
because the DNS resolving is performed by GThreadedResolver, which uses
a GTask taking a strong reference to the agent main context in order to
be able finish the task and run the final callback within the agent main
context. Consequently, the main context cannot be disposed until the
GTask has finished. Leaking a GMainContext is particularly bad because
it involves leaking the file descriptor used for polling events.

GThreadedResolver takes a strong reference to the GTask. it hands this
reference over to an internal worker thread that ultimately calls libc's
`getaddrinfo()`. This function can hang for an undefined amount of time
depending on network or network driver conditions. This makes it
impossible to completely control the lifetime of our GTask/GMainContext,
but what we can do is to make sure that in case of a long hang or early
shutdown of the agent, the GTask has already finished so that when
`getaddrinfo()` returns the only work left to do is to drop the
reference to GTask (and possibly GMainContext).

GThreadedResolver already sets up two mechanisms for finishing the task
earlier than `getaddrinfo()`, one timeout source and a cancellation
source. The cancellation source was previously created by
GThreadedResolver, but we're now passing it ourselves so that we can
explicitly invoke cancellation. There is one quirk though. Both the
timeout source and the cancellation source are associated with the gio
global worker context. Thus, cancellation will happen in the thread
driving the worker context and there it will conclude that the final
agent callback must be invoked within the agent main context. This
detour means we cannot simply cancel and expect the GTask to be
finalized without also explicitly waiting for the finalization.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/8519>