lcevcdecodebin: Update the base decoder when setting base-decoder property
Currently, the base-decoder property only works when setting it while
constructing the element, but does not work if we set the property after
constructing the element. This patch fixes this issue so that the property
can be set with gst-launch-1.0. Note that the property can only be set if
the element is in NULL state.

Fixes #4594

Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/9599>