wayland: window: Delay the render rectangle once the surface is configured
Updating the surface before sending the configure ack results in bad window
placement. Handle this situation like if the window is being created by
unsetting the configured flag.

Once we acked the configuration, always update the geometry, which ensure we
have a commit after that event regardless if anything dimensions changed or not.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/9653>