dwrite: Fix D3D12 critical warning
When using convert + blend mode, the final render target resource
is also used as a source resource owned by our fence data storage.
As a result, the resource may not be writable if the conversion
command has not yet finished. To address this, add a private
method that allows setting the fence without a writability check

Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/9681>