wayland: window: Protect against too small window
Fix protocol error due to 0 sized destination. When the compositor decides we
should render to a really small window such as 1x1, the inner rectangle for the
video surface may endup with a width or height of 0, which is a protocol error.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/9653>