gsth264parser: fix uint32 to int32 truncation
The H264 parser is using int32 to compute the framerate numerator and
denominator while the SPS block holding the original values used for
this computation are stored on uint32. On corner cases it may lead to a
truncation of the final values leading to an invalid framerate.

This patch is shifting right the numerator and denominator until both
values may be stored on int32. It may introduce a small computation
error with odd values but negligeable taking into account the huge
initial values > max_int32.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/9967>