vkh264enc: remove unused member variable
The member variable out_state is not used in the code, so there's no need to
keep it as a member variable along the lifespan of the object.

Even more, it can be fetched with gst_video_encoder_get_output_state().

Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/10080>